Refactor additional shared methods into mixin
