+++ coldfront/core/project/tests/test_utils/test_renewal_utils/test_allocation_renewal_denial_runner.py
@@ -13,32 +13,6 @@ from django.test import TestCase
 class TestRunnerMixin(TestRunnerMixinBase):
     """A mixin for testing AllocationRenewalDenialRunner."""
 
-    def assert_pooling_preference_case(self, expected):
-        """Assert that the pooling preference case of the request_obj is
-        the provided, expected one."""
-        actual = self.request_obj.get_pooling_preference_case()
-        self.assertEqual(expected, actual)
-
-    def create_request(self, pi=None, pre_project=None, post_project=None,
-                       new_project_request=None):
-        """Create and return an AllocationRenewalRequest with the given
-        parameters."""
-        assert pi and pre_project and post_project
-        approved_renewal_request_status = \
-            AllocationRenewalRequestStatusChoice.objects.get(
-                name='Under Review')
-        kwargs = {
-            'requester': self.requester,
-            'pi': pi,
-            'allocation_period': self.allocation_period,
-            'status': approved_renewal_request_status,
-            'pre_project': pre_project,
-            'post_project': post_project,
-            'request_time': utc_now_offset_aware(),
-            'new_project_request': new_project_request,
-        }
-        return AllocationRenewalRequest.objects.create(**kwargs)
-
     def test_request_initial_not_complete_status_enforced(self):
         """Test that the provided AllocationRenewalRequest must not be
         in the 'Complete' state, or an exception will be raised."""
@@ -149,6 +123,8 @@ class TestUnpooledToUnpooled(TestRunnerMixin, TestCase):
         """Set up test data."""
         super().setUp()
         self.request_obj = self.create_request(
+            AllocationRenewalRequestStatusChoice.objects.get(
+                name='Under Review'),
             pi=self.pi0,
             pre_project=self.unpooled_project0,
             post_project=self.unpooled_project0)
@@ -168,6 +144,8 @@ class TestUnpooledToPooled(TestRunnerMixin, TestCase):
         """Set up test data."""
         super().setUp()
         self.request_obj = self.create_request(
+            AllocationRenewalRequestStatusChoice.objects.get(
+                name='Under Review'),
             pi=self.pi0,
             pre_project=self.unpooled_project0,
             post_project=self.pooled_project1)
@@ -187,6 +165,8 @@ class TestPooledToPooledSame(TestRunnerMixin, TestCase):
         """Set up test data."""
         super().setUp()
         self.request_obj = self.create_request(
+            AllocationRenewalRequestStatusChoice.objects.get(
+                name='Under Review'),
             pi=self.pi0,
             pre_project=self.pooled_project0,
             post_project=self.pooled_project0)
@@ -206,6 +186,8 @@ class TestPooledToPooledDifferent(TestRunnerMixin, TestCase):
         """Set up test data."""
         super().setUp()
         self.request_obj = self.create_request(
+            AllocationRenewalRequestStatusChoice.objects.get(
+                name='Under Review'),
             pi=self.pi0,
             pre_project=self.pooled_project0,
             post_project=self.pooled_project1)
@@ -225,6 +207,8 @@ class TestPooledToUnpooledOld(TestRunnerMixin, TestCase):
         """Set up test data."""
         super().setUp()
         self.request_obj = self.create_request(
+            AllocationRenewalRequestStatusChoice.objects.get(
+                name='Under Review'),
             pi=self.pi0,
             pre_project=self.pooled_project0,
             post_project=self.unpooled_project0)
@@ -246,6 +230,8 @@ class TestPooledToUnpooledNew(TestNewProjectDenialMixin, TestRunnerMixin,
         super().setUp()
         new_project_request = self.create_under_review_new_project_request()
         self.request_obj = self.create_request(
+            AllocationRenewalRequestStatusChoice.objects.get(
+                name='Under Review'),
             pi=self.pi0,
             pre_project=self.pooled_project0,
             post_project=new_project_request.project,
